SIGMA: Node Manager & More
	Implement an advanced Node manager into the Scene class
	to handle Node Tree changes safely.
	This system was moved from Node.cpp, and improved.

	Other minor changes include:
		- GUI system tweaks
		- LightNode.cpp cleanup (this class is not needed yet, but I'll save it for later)
		- Time double -> float
		- Time Scale
		- Input Cursor ( Now you can change the cursor! )

+++ Time.cpp
@@ -26,7 +26,7 @@ void sig::Time::Release()
 	ms_instance = nullptr;
 }
 
-double sig::Time::GetTime()
+float sig::Time::GetTime()
 {
-	return SDL_GetTicks() / 1000.0;
+	return float(SDL_GetTicks()) / 1000.0f;
 }
